refactor: replace use of localStorage w/ Firebase methods
Since the use of localStorage to persist user tokens is a possible
security issue, and it makes this information difficult to delete if the
user so wishes, all instances of storing and retrieving the Firebase
token in localStorage have been replaced with native Firebase methods.
Namely, `auth.currentUser` to get the current user object, and the
`user.getIdToken` method to get the token itself.